[Submarine-339] Separate SubmarineConfVars from SubmarineConfiguration
### What is this PR for?
SubmarineConfiguration becomes larger and larger. Hence, to facilitate the management of configuration items, I separate the class SubmarineConfVars from the class SubmarineConfiguration.

### What type of PR is it?
[Refactoring]

### Todos

### What is the Jira issue?
https://issues.apache.org/jira/browse/SUBMARINE-339

### How should this be tested?
It will be tested by Travis.

### Screenshots (if appropriate)
https://travis-ci.org/apache/submarine/builds/636675316?utm_source=github_status&utm_medium=notification

### Questions:
* Does the licenses files need to be updated? No
* Are there breaking changes for older versions? No
* Does this needs documentation? No

Author: kevin85421 <b03901153@ntu.edu.tw>

Closes #151 from kevin85421/SUBMARINE-339 and squashes the following commits:

885d409 [kevin85421] [SUBMARINE-339] Separate SubmarineConfVars from SubmarineConfiguration
e37765d [kevin85421] [SUBMARINE-339] Separate SubmarineConfVars from SubmarineConfiguration
7306645 [kevin85421] [SUBMARINE-339] Separate SubmarineConfVars from SubmarineConfiguration
