+++ submarine-commons/commons-runtime/src/main/java/org/apache/submarine/commons/runtime/RuntimeFactory.java
@@ -21,6 +21,7 @@ package org.apache.submarine.commons.runtime;
 
 import com.google.common.annotations.VisibleForTesting;
 import org.apache.submarine.commons.utils.SubmarineConfiguration;
+import org.apache.submarine.commons.utils.SubmarineConfVars;
 import org.apache.submarine.commons.runtime.exception.SubmarineRuntimeException;
 import org.apache.submarine.commons.runtime.fs.SubmarineStorage;
 
@@ -41,7 +42,7 @@ public abstract class RuntimeFactory {
     SubmarineConfiguration submarineConfiguration =
         clientContext.getSubmarineConfig();
     String runtimeClass = submarineConfiguration.getString(
-        SubmarineConfiguration.ConfVars.SUBMARINE_RUNTIME_CLASS);
+        SubmarineConfVars.ConfVars.SUBMARINE_RUNTIME_CLASS);
 
     try {
       Class<?> runtimeClazz = Class.forName(runtimeClass);
